book_image intergation with cloudinary

+++ routes/admin.js
@@ -1,19 +1,17 @@
 const express = require("express");
 const pool = require("../db/db");
 const multer = require("multer");
-const crypto = require("crypto");
+const path = require("path");
+const cloudinary = require("cloudinary").v2;
 
-const storage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, "./images");
-  },
-  filename: function (req, file, cb) {
-    cb(null, crypto.randomBytes(10).toString("hex") + file.originalname);
-  },
+cloudinary.config({
+  cloud_name: process.env.CLOUD_NAME,
+  api_key: process.env.API_KEY,
+  api_secret: process.env.API_SECRET,
+  secure: true,
 });
 
 const fileFilter = (req, file, cb) => {
-  // reject a file
   if (file.mimetype === "image/jpeg" || file.mimetype === "image/png") {
     cb(null, true);
   } else {
@@ -22,7 +20,7 @@ const fileFilter = (req, file, cb) => {
 };
 
 const upload = multer({
-  storage: storage,
+  storage: multer.diskStorage({}),
   limits: {
     fileSize: 1024 * 1024 * 5,
   },
@@ -48,18 +46,25 @@ route.get("/upload/book", (req, res) => {
   res.render("books");
 });
 
-route.post("/upload/book", upload.single("book_image"), (req, res) => {
-  const { name, author, price, description } = req.body;
-  const image = req.file.filename;
-  const book_image = `http://${req.headers.host}/images/${image}`;
-  pool.query(
-    `INSERT INTO books(author,price,name,description,book_image) VALUES($1,$2,$3,$4,$5) RETURNING  author,price,name,description,book_image`,
-    [author, price, name, description, book_image],
-    (err, results) => {
-      if (err) throw err;
-      res.render("admin", { book: results.rows });
-    }
-  );
+route.post("/upload/book", upload.single("book_image"), async (req, res) => {
+  try {
+    const result = await cloudinary.uploader.upload(req.file.path, {
+      folder: "bookstore",
+    });
+    const { name, author, price, description } = req.body;
+    const image = req.file.filename;
+    const book_image = result.secure_url;
+    pool.query(
+      `INSERT INTO books(author,price,name,description,book_image) VALUES($1,$2,$3,$4,$5) RETURNING  author,price,name,description,book_image`,
+      [author, price, name, description, book_image],
+      (err, results) => {
+        if (err) throw err;
+        res.render("admin", { book: results.rows });
+      }
+    );
+  } catch {
+    res.status(400).send("Error");
+  }
 });
 
 route.get("/getbooks", (req, res) => {
@@ -85,10 +90,21 @@ route.post("/updatebook", async (req, res) => {
 
 route.post("/deletebook", async (req, res) => {
   const { id } = req.body;
-
-  await pool.query(`DELETE FROM books WHERE id = $1`, [id], (err, results) => {
+  await pool.query(`SELECT * FROM books WHERE id=$1`, [id], (err, results) => {
     if (err) throw err;
-    res.redirect("/admin/getbooks");
+    if (results.rows.length > 0) {
+      try {
+        let url = results.rows[0].book_image.split("/");
+        var lastsegment = url[url.length - 1].split(".");
+        pool.query(`DELETE FROM books WHERE id = $1`, [id], (err, results) => {
+          if (err) throw err;
+          cloudinary.uploader.destroy(`bookstore/${lastsegment[0]}`);
+          res.redirect("/admin/getbooks");
+        });
+      } catch {
+        res.status(400).redirect("/admin");
+      }
+    }
   });
 });
